Conflicts:
	src/1. Layers/1.1 Presentation/Well.Dashboard/app/delivery/model/delivery.ts
	src/1. Layers/1.1 Presentation/Well.Dashboard/app/delivery/model/deliveryLine.ts
	src/1. Layers/1.3 Domain/Well.Domain/ValueObjects/Delivery.cs
	src/3. Database/Well/dbo/StoredProcedures/Job_GetByIds.sql

+++ src/3. Database/Well/dbo/Tables/Job.sql
@@ -28,7 +28,7 @@
 	[IsDeleted] BIT NOT NULL DEFAULT 0,
 	[StopId] INT NOT NULL,
 	[ActionLogNumber] VARCHAR(50) NULL,
-	[OuterCount] TINYINT NULL,
+	[OuterCount] INT NULL,
 	[OuterDiscrepancyFound] BIT DEFAULT 0,
 	[TotalOutersOver] INT NULL,
 	[TotalOutersShort] INT NULL,
